Add files via upload
This code is for editing the contents of a cell in an Excel file. This code can edit excel in large numbers at once, I made this to make my freelance work easier. but I was confused when I installed it on my friend's laptop, this .exe was detected as a trojan. Please correct me, is there anything wrong with the code? pleaseee help me